+++ app.js
@@ -10,10 +10,10 @@ var logger = require("morgan");
 var indexRouter = require("./routes/index");
 var usersRouter = require("./routes/users");
 var loginRouter = require("./routes/member/login");
-var searchRouter = require("./routes/search");
-var tagRouter = require("./routes/tag");
+var searchRouter = require("./routes/study/search");
+var tagRouter = require("./routes/study/tag");
 var listRouter = require("./routes/study/list");
-var studyRouter = require("./routes/study");
+var studyRouter = require("./routes/study/study");
 
 var app = express();
 
@@ -43,9 +43,9 @@ app.use(express.static(path.join(__dirname, "public")));
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/login", loginRouter);
-app.use("/search", searchRouter);
+app.use("/study/search", searchRouter);
 app.use("/tag", tagRouter); // localhost:3000/tag
-app.use("/list", listRouter);
+app.use("/study/list", listRouter);
 app.use("/study", studyRouter);
 
 app.use(function (req, res, next) {
